+++ lib/src/commands/commands.dart
@@ -1,3 +1,4 @@
+export 'change_app_name_command.dart';
 export 'change_id_command.dart';
 export 'create_command.dart';
 export 'feature_command.dart';